💚 Fix: solve the mybatis query problem (Cannot find the dto.. cause mix file structure
파일 구조 변화로 인해 dto 못 찾아서 에러 뜨는 문제 해결
