+++ src/main/java/com/handongapp/cms/mapper/ClubMapper.java
@@ -1,6 +1,6 @@
 package com.handongapp.cms.mapper;
 
-import com.handongapp.cms.dto.TbClubDto;
+import com.handongapp.cms.dto.v1.TbClubDto;
 import org.apache.ibatis.annotations.Param;
 
 public interface ClubMapper {
