+++ src/main/java/com/handongapp/cms/repository/TbCourseRepositoryCustom.java
@@ -1,0 +1,9 @@
+package com.handongapp.cms.repository;
+
+import com.handongapp.cms.dto.v1.TbClubDto;
+
+import java.util.List;
+
+public interface TbCourseRepositoryCustom {
+    List<TbClubDto.ClubCourseListResDto> findCoursesByClubIdAndProgramId(String clubName, String programId);
+}
