+++ src/main/java/com/handongapp/cms/repository/impl/TbCourseRepositoryImpl.java
@@ -1,19 +1,19 @@
-package com.handongapp.cms.repository.Impl;
+package com.handongapp.cms.repository.impl;
 
 import com.handongapp.cms.domain.*;
-import com.handongapp.cms.dto.TbClubDto;
-import com.handongapp.cms.repository.Custom.CourseRepositoryCustom;
+import com.handongapp.cms.dto.v1.TbClubDto;
+import com.handongapp.cms.repository.TbCourseRepositoryCustom;
 
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
 import java.util.List;
 
-public class CourseRepositoryImpl implements CourseRepositoryCustom {
+public class TbCourseRepositoryImpl implements TbCourseRepositoryCustom {
 
     private final JPAQueryFactory queryFactory;
 
-    public CourseRepositoryImpl(JPAQueryFactory queryFactory) {
+    public TbCourseRepositoryImpl(JPAQueryFactory queryFactory) {
         this.queryFactory = queryFactory;
     }
 
@@ -23,7 +23,7 @@ public class CourseRepositoryImpl implements CourseRepositoryCustom {
     // programId로 tb_course랑 연결해서 picture_url 받아오기
 
     @Override
-    public List<TbClubDto.ClubCourseListResDto> findCoursesByClubAndProgram(String clubName, String programId) {
+    public List<TbClubDto.ClubCourseListResDto> findCoursesByClubIdAndProgramId(String clubName, String programId) {
         QTbClub club = QTbClub.tbClub;
         QTbProgram program = QTbProgram.tbProgram;
         QTbUser user = QTbUser.tbUser;
