+++ src/main/java/com/handongapp/cms/service/CourseService.java
@@ -1,11 +1,8 @@
 package com.handongapp.cms.service;
 
-import com.handongapp.cms.dto.TbClubDto;
-import com.handongapp.cms.repository.TbCourseRepository;
-import org.springframework.stereotype.Service;
+import com.handongapp.cms.dto.v1.TbClubDto;
 
 import java.util.List;
-
 
 public interface CourseService {
     List<TbClubDto.ClubCourseListResDto> getCourseList(String clubName, String programId);
