+++ src/main/java/com/handongapp/cms/service/TbUserService.java
@@ -2,7 +2,7 @@ package com.handongapp.cms.service;
 
 import com.handongapp.cms.auth.dto.GoogleUserInfoResponse;
 import com.handongapp.cms.domain.TbUser;
-import com.handongapp.cms.dto.TbUserDto;
+import com.handongapp.cms.dto.v1.TbUserDto;
 
 public interface TbUserService {
     TbUser saveOrUpdateUser(String userId, String email, String name);
